Fixed wrong line number output in the error message box

+++ Editor/mainwindow.h
@@ -186,7 +186,7 @@ public:
         /// \param lineNum number Number of line where error found. If <0 - don't show line information.
         /// \param lineContents contents of line with error. If empty string - don't show field.
         ///
-        void formatErrorMsgBox(QString filePath, QString errorMessage, int lineNum=-1, QString lineContents="");
+        void formatErrorMsgBox(QString filePath, QString errorMessage, long lineNum=-1, QString lineContents="");
         ///
         /// \brief setDefaults Init settings on start application
         ///
